+++ src/fresnel_transform/rss_ringoccs_fresnel_transform_newton.c
@@ -77,13 +77,11 @@
  *          The diffraction corrected profile.                                *
  ******************************************************************************/
 void
-rssringoccs_Fresnel_Transform_Newton(rssringoccs_TAUObj *tau,
-                                     double *w_func,
-                                     unsigned long n_pts,
-                                     unsigned long center)
+rssringoccs_Fresnel_Transform_Newton(rssringoccs_TAUObj *tau, double *w_func,
+                                     size_t n_pts, size_t center)
 {
     /*  Declare all necessary variables. i and j are used for indexing.       */
-    unsigned long m, offset;
+    size_t m, offset;
 
     /*  The Fresnel kernel and ring azimuth angle.                            */
     double psi, phi, cos_psi, sin_psi, factor;
